refactor: move private cli code in internal folder
